+++ .gitignore
@@ -2,3 +2,4 @@
 /.idea/
 /gradle/
 /build/
+middlejava.iml
