--- src/main/java/com/application/controller/QuizController.java
@@ -1,12 +1,0 @@
-package com.application.controller;
-
-import com.application.service.QuizService;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Controller;
-
-@Controller
-@RequiredArgsConstructor
-public class QuizController {
-
-    private final QuizService quizService;
-}
