--- src/main/java/com/application/controller/UserController.java
@@ -1,12 +1,0 @@
-package com.application.controller;
-
-import com.application.service.UserService;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Controller;
-
-@Controller
-@RequiredArgsConstructor
-public class UserController {
-
-    private final UserService userService;
-}
