--- src/main/java/com/application/dto/create/UserCreateDto.java
@@ -1,4 +1,0 @@
-package com.application.dto.create;
-
-public class UserCreateDto {
-}
