--- src/main/java/com/application/dto/get/QuestionGetDto.java
@@ -1,4 +1,0 @@
-package com.application.dto.get;
-
-public class QuestionGetDto {
-}
