--- src/main/java/com/application/dto/get/QuizGetDto.java
@@ -1,4 +1,0 @@
-package com.application.dto.get;
-
-public class QuizGetDto {
-}
