--- src/main/java/com/application/dto/get/UserGetDto.java
@@ -1,4 +1,0 @@
-package com.application.dto.get;
-
-public class UserGetDto {
-}
