--- src/main/java/com/application/dto/update/QuizUpdateDto.java
@@ -1,4 +1,0 @@
-package com.application.dto.update;
-
-public class QuizUpdateDto {
-}
