--- src/main/java/com/application/dto/update/UserUpdateDto.java
@@ -1,4 +1,0 @@
-package com.application.dto.update;
-
-public class UserUpdateDto {
-}
