--- src/main/java/com/application/model/BaseEntity.java
@@ -1,18 +1,0 @@
-package com.application.model;
-
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.MappedSuperclass;
-import lombok.Getter;
-import lombok.Setter;
-
-@MappedSuperclass
-@Getter
-@Setter
-public abstract class BaseEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long id;
-}
