--- src/main/java/com/application/model/Role.java
@@ -1,12 +1,0 @@
-package com.application.model;
-
-import org.springframework.security.core.GrantedAuthority;
-
-public enum Role implements GrantedAuthority {
-    ROLE_USER,
-    ROLE_ADMIN;
-
-    public String getAuthority() {
-        return name();
-    }
-}
