--- src/main/java/com/application/repository/AnswerRepository.java
@@ -1,9 +1,0 @@
-package com.application.repository;
-
-import com.application.model.Answer;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface AnswerRepository extends JpaRepository<Answer, Long> {
-}
