--- src/main/java/com/application/repository/QuestionRepository.java
@@ -1,9 +1,0 @@
-package com.application.repository;
-
-import com.application.model.Question;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface QuestionRepository extends JpaRepository<Question, Long> {
-}
