--- src/main/java/com/application/repository/QuizRepository.java
@@ -1,9 +1,0 @@
-package com.application.repository;
-
-import com.application.model.Quiz;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface QuizRepository extends JpaRepository<Quiz, Long> {
-}
