--- src/main/java/com/application/repository/UserRepository.java
@@ -1,9 +1,0 @@
-package com.application.repository;
-
-import com.application.model.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface UserRepository extends JpaRepository<User, Long> {
-}
