--- src/main/java/com/application/service/AnswerService.java
@@ -1,12 +1,0 @@
-package com.application.service;
-
-import com.application.repository.AnswerRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-
-@Service
-@RequiredArgsConstructor
-public class AnswerService {
-
-    private final AnswerRepository answerRepository;
-}
