--- src/main/java/com/application/service/QuestionService.java
@@ -1,12 +1,0 @@
-package com.application.service;
-
-import com.application.repository.QuestionRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-
-@Service
-@RequiredArgsConstructor
-public class QuestionService {
-
-    private final QuestionRepository questionRepository;
-}
