--- src/main/java/com/application/service/QuizService.java
@@ -1,12 +1,0 @@
-package com.application.service;
-
-import com.application.repository.QuizRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-
-@Service
-@RequiredArgsConstructor
-public class QuizService {
-
-    private final QuizRepository quizRepository;
-}
