--- src/main/java/com/application/service/UserService.java
@@ -1,12 +1,0 @@
-package com.application.service;
-
-import com.application.repository.UserRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-
-@Service
-@RequiredArgsConstructor
-public class UserService {
-
-    private final UserRepository userRepository;
-}
