+++ src/main/java/com/quiz/controller/AnswerController.java
@@ -1,0 +1,12 @@
+package com.quiz.controller;
+
+import com.quiz.service.AnswerService;
+import lombok.RequiredArgsConstructor;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequiredArgsConstructor
+public class AnswerController {
+
+    private final AnswerService answerService;
+}
