+++ src/main/java/com/quiz/controller/QuestionController.java
@@ -1,0 +1,12 @@
+package com.quiz.controller;
+
+import com.quiz.service.QuestionService;
+import lombok.RequiredArgsConstructor;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequiredArgsConstructor
+public class QuestionController {
+
+    private final QuestionService questionService;
+}
