+++ src/main/java/com/quiz/controller/QuizController.java
@@ -1,0 +1,12 @@
+package com.quiz.controller;
+
+import com.quiz.service.QuizService;
+import lombok.RequiredArgsConstructor;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequiredArgsConstructor
+public class QuizController {
+
+    private final QuizService quizService;
+}
