+++ src/main/java/com/quiz/controller/UserController.java
@@ -1,0 +1,12 @@
+package com.quiz.controller;
+
+import com.quiz.service.UserService;
+import lombok.RequiredArgsConstructor;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequiredArgsConstructor
+public class UserController {
+
+    private final UserService userService;
+}
