+++ src/main/java/com/quiz/dto/AnswerDto.java
@@ -1,0 +1,4 @@
+package com.quiz.dto;
+
+public class AnswerDto {
+}
