+++ src/main/java/com/quiz/dto/QuestionDto.java
@@ -1,0 +1,4 @@
+package com.quiz.dto;
+
+public class QuestionDto {
+}
