+++ src/main/java/com/quiz/dto/UserDto.java
@@ -1,0 +1,4 @@
+package com.quiz.dto;
+
+public class UserDto {
+}
