+++ src/main/java/com/quiz/model/Answer.java
@@ -1,0 +1,22 @@
+package com.quiz.model;
+
+import javax.persistence.*;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.Setter;
+
+@Entity
+@Builder
+@Getter
+@Setter
+@Table(name = "answers")
+public class Answer {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private long id;
+
+    private String answerValue;
+
+    private boolean isCorrect;
+}
