+++ src/main/java/com/quiz/model/Quiz.java
@@ -1,0 +1,23 @@
+package com.quiz.model;
+
+import java.util.List;
+import javax.persistence.*;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.Setter;
+
+@Entity
+@Builder
+@Getter
+@Setter
+@Table(name = "quizzes")
+public class Quiz {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private long id;
+
+    private String quizName;
+
+    private List<Question> questions;
+}
