+++ src/main/java/com/quiz/model/Role.java
@@ -1,0 +1,12 @@
+package com.quiz.model;
+
+import org.springframework.security.core.GrantedAuthority;
+
+public enum Role implements GrantedAuthority {
+    ROLE_USER,
+    ROLE_ADMIN;
+
+    public String getAuthority() {
+        return name();
+    }
+}
