+++ src/main/java/com/quiz/model/User.java
@@ -1,0 +1,28 @@
+package com.quiz.model;
+
+import java.util.Set;
+import javax.persistence.*;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.Setter;
+
+@Entity
+@Builder
+@Getter
+@Setter
+@Table(name = "users")
+public class User {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private long id;
+
+    @Column
+    private String name;
+
+    @Column
+    private String lastName;
+
+    @Column(name = "role")
+    private Set<Role> roles;
+}
