+++ src/main/java/com/quiz/repository/AnswerRepository.java
@@ -1,0 +1,9 @@
+package com.quiz.repository;
+
+import com.quiz.model.Answer;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface AnswerRepository extends JpaRepository<Answer, Long> {
+}
