+++ src/main/java/com/quiz/repository/QuestionRepository.java
@@ -1,0 +1,9 @@
+package com.quiz.repository;
+
+import com.quiz.model.Question;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface QuestionRepository extends JpaRepository<Question, Long> {
+}
