+++ src/main/java/com/quiz/repository/QuizRepository.java
@@ -1,0 +1,9 @@
+package com.quiz.repository;
+
+import com.quiz.model.Quiz;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface QuizRepository extends JpaRepository<Quiz, Long> {
+}
