+++ src/main/java/com/quiz/repository/UserRepository.java
@@ -1,0 +1,9 @@
+package com.quiz.repository;
+
+import com.quiz.model.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface UserRepository extends JpaRepository<User, Long> {
+}
