+++ src/main/java/com/quiz/service/AnswerService.java
@@ -1,0 +1,12 @@
+package com.quiz.service;
+
+import com.quiz.repository.AnswerRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+public class AnswerService {
+
+    private final AnswerRepository answerRepository;
+}
