+++ src/main/java/com/quiz/service/QuestionService.java
@@ -1,0 +1,12 @@
+package com.quiz.service;
+
+import com.quiz.repository.QuestionRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+public class QuestionService {
+
+    private final QuestionRepository questionRepository;
+}
