+++ src/main/java/com/quiz/service/QuizService.java
@@ -1,0 +1,12 @@
+package com.quiz.service;
+
+import com.quiz.repository.QuizRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+public class QuizService {
+
+    private final QuizRepository quizRepository;
+}
