+++ src/main/java/com/quiz/service/UserService.java
@@ -1,0 +1,12 @@
+package com.quiz.service;
+
+import com.quiz.repository.UserRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+public class UserService {
+
+    private final UserRepository userRepository;
+}
